smell library
This is the smell library source code which is being used in smellputer
project. The library is used to save, edit, insert smell at a specific
pin on the device as well as it can be used to attach it with the image
if its available. The library stores smell with their properties and
pin they are present on in an xml file whose reference is later used to
attach the smell with the image.